+++ docs/plans/2026-08-16-review-and-autoclose-prs-329-through-342.md
@@ -27,6 +27,8 @@ Linked plans:
   Turbo/CI input cleanup and throwable scaffold command failures.
 - [PR #334 autoclosure](docs/plans/2026-08-16-pr-334-autoclosure.md) -
   Example/docs render and read-path cleanup with browser proof.
+- [PR #332 autoclosure](docs/plans/2026-08-16-pr-332-autoclosure.md) -
+  React identity stability and auth token rotation.
 
 Completion threshold:
 - All 14 frozen PRs (#329-#342) have a recorded intent, overlap/dependency map,
@@ -108,9 +110,9 @@ PR disposition ledger:
 | #329 | Solid infinite query mounts against real Solid Query | not slop; real compact correctness fix | 0 threads | first; #339 depends on it | closed: merged/released 0.17.2, post-release CI green |
 | #330 | lazy CLI startup and bounded codegen rework | credible but broad; not slop | repaired: 1 outdated + 1 current P2 | before #337/#342 | closed: merged `b78d94a5`, remote gates green |
 | #331 | remove ORM write-path rereads without corrupting hooks/RLS | credible but oversized; not slop | 1 obsolete + 1 repaired P1; structured review found/fixed another RLS P1 | before #335/#336/#337/#342 | closed: merged `23c0c999`, remote gates green |
-| #332 | preserve React queries across token rotation | credible identity fix | 1 outdated changeset thread | before #339 | verify pending |
+| #332 | preserve React queries across token rotation | credible but overstuffed React identity work; stale release story is slop | 1 outdated changeset thread already fixed to patch | before #339 | current main merged; focused proof green |
 | #333 | correct Turbo/CI inputs and delete dead work | credible tooling cleanup | 0 threads | before #342; ride 0.18 release | closed: merged `413d4068`, remote gates green |
-| #334 | reduce example/www client work and reads | mixed but mostly real; three proof/owner defects were slop and repaired | browser P1 + dedicated-existence P2 | package codegen fix joins pending patch release | local proof complete; delivery pending |
+| #334 | reduce example/www client work and reads | mixed but mostly real; three proof/owner defects were slop and repaired | both threads resolved | codegen fix released in 0.17.4 | closed: merged `6bc8d1ca`, released 0.17.4 |
 | #335 | push ORM limit/order into indexes | valuable but high-risk | 3 P1 threads | before #336/#337/#342 | repair pending |
 | #336 | bound streams and id-list pagination | credible, breaking | 2 P2 threads | after #335 | repair pending |
 | #337 | bound aggregate reads and clearing | valuable but high-risk | 2 P1 threads | before #342 | repair pending |
@@ -122,8 +124,8 @@ PR disposition ledger:
 
 Dependency-safe order:
 - Release 0.17.2: #329.
-- Pending patch release: #331, #333, #334 after proof.
-- Next release trigger: #332 after the pending patch release settles.
+- Released 0.17.4: #331 plus #334's package repair; #333 had no package delta.
+- Next patch release: #332.
 - Release 0.19.x: #335 then #336.
 - Release 0.20.x: #337 then #338.
 - Release 0.21.x: #339 then #340.
@@ -223,6 +225,13 @@ Verification evidence:
   regenerate 17 routes, and pass 4 focused read-count tests, 63 codegen tests,
   package build, example/Convex typechecks, five Browser routes, responsive
   walkthrough proof, zero slop delta, lint, and full `bun check`.
+- #334 head `26b8f4de` passed CI `31971886260` and Vercel, then merged as
+  `6bc8d1ca`; release PR #346 passed the 15-minute Convex Matrix/runtime lane,
+  merged as `e79a0d8d`, and published npm/GitHub `v0.17.4`.
+- #332 merged released main cleanly. Its 80 focused React tests pass. The auth
+  identity comparison and referential-stability invariants are real; its stale
+  claim that an additive export requires a minor is release-story slop. The
+  branch itself already follows repository v0 policy with a patch changeset.
 
 Timeline:
 - 2026-08-16T18:31:28.829Z Autoclosure plan created.
